Removed noise reference
