+++ src/chap2.qmd
@@ -58,13 +58,7 @@ Comparisons may be random and are generated by random draws of (total) orders. (
  
 In the following, we consider randomness as generated from a *decision-maker* who has an order, or preference relation, $\prec$ on a set of objects $Y$. We refer to the random object $\prec$ as the oracle preference. Each preference $\prec$ has an associated probability mass $\mathbb{P}[\mathord{\prec}]$, leading to an $(n!-1)$-dimensional vector encoding the full random preference set. (This might look like, and is already for small values of $n$ a large number. Reducing this representational complexity is a goal of this chapter.)
 
-One might wonder why we need to have a random preference. Deterministic preferences are conceptually helpful constructs and are used broadly in the fields of Consumer theory (e.g., @mas1995microeconomic). However, they suffer when bringing them to data, as data is inherently noisy. 
-
-Noise plays three roles, which we will address in the chapters of this book. 
-
- - It first captures measurement noise. When observing choices from a person, this may be an imperfect preference relationship. We will consider such noise (homogeneity but uncertainty), for example, in @sec-learning when discussing Gaussian processes.
- - 
- - A third role of noise arises from heterogeneity, where $\prec$ encodes the *population* preferences of several decision-makers. We consider an example like this in one of the exercises in this chapter.
+One might wonder why we need to have a random preference. Deterministic preferences are conceptually helpful constructs and are used broadly in the fields of Consumer theory (e.g., @mas1995microeconomic). However, they suffer when bringing them to data, as data is inherently noisy.
 
 Even when allowing for randomness, assumptions we will impose in this section, be it transitivity or the Independence of Irrelevant Alternatives, are stark yet practical. In many situations they will fail, for good reasons. Whether these are human's inability to express rankings, contextual challenges of domains, or community norms, we will discuss them in, humans cannot clearly rank alternatives, their choices reflect individualistic norms, or they might have self-control pictures. Many of these wrinkles on the approach to preferences presented here is contained in @sec-beyond. Until then, we will make the fullest use of learning stochastic preferences.
